+++ src/tab/formatter/formatMarkdown.js
@@ -1,4 +1,4 @@
-import {getUrl} from "../utils/webpage/getters";
+import {getUrl} from "../utils/getters";
 import {capitalizeFirst, formatDate, formatLineBreaks, titleShortener} from "./formatText";
 import TurndownService from "turndown";
 import DOMPurify from "dompurify";
