+++ src/tab/scraper/scraper.js
@@ -3,7 +3,7 @@ import {logWelcome} from "../utils/consoleMessages";
 import {setFormatRules} from "./ruler/ruler";
 import {extract} from "./extractor/extractor";
 import {exportContent} from "./exporter/exporter";
-import {clickOnListElt} from "../utils/webpage/interact";
+import {clickOnListElt} from "../uiEnhancer/phind/interact";
 
 /**
  * @description - Launch the export process
